requirements.txt
pytest~=8.4.0
selene==2.0.0rc9
allure-pytest
allure-python-commons~=2.14.3